+++ CHANGELOG.md
@@ -266,3 +266,8 @@
 ### v1.0 9/26/19
 * added Wifite2 to network_crack.py
 * added vBulletin 5-5.5.4 Scan to domain_sticate.py and vBulletinScan
+
+### v1.0 9/28/19
+* added fluxion to network_crack.py - for real this time
+* added ettercap to network_crack.py and lib_install.sh
+* added necessary libs for running ettercap